백준 11382

+++ src/_001/_010.java
@@ -12,9 +12,9 @@ public class _010 {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         String[] input = br.readLine().split(" ");
-        long a = Integer.parseInt(input[0]);
-        long b = Integer.parseInt(input[1]);
-        long c = Integer.parseInt(input[2]);
+        long a = Long.parseLong(input[0]);
+        long b = Long.parseLong(input[1]);
+        long c = Long.parseLong(input[2]);
         System.out.println(a+b+c);
     }
 }
